Fixed majority of Xcode warnings - Converted all int to NSInteger to ensure cross processor architecture compatibility - Added missing CAAnimationDelegate specifiers - Added correct casting for integers in format strings

+++ Classes/Chapter.h
@@ -10,10 +10,10 @@
 
 @interface Chapter : NSObject
 
-@property (nonatomic) int chapterNumber;
+@property (nonatomic) NSInteger chapterNumber;
 @property (nonatomic, copy) NSString *title;
 @property (nonatomic, copy) NSArray *lessons;
 
--(id)initWithChapterNumber:(int)chapterNumber;
+-(id)initWithChapterNumber:(NSInteger)chapterNumber;
 
 @end
